requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
redis>=5.0.0
python-multipart>=0.0.9
sse-starlette>=2.0.0
tiktoken>=0.5.0

# LLM Providers (optional, uncomment as needed)
# openai>=1.10.0
# anthropic>=0.18.0

# Vector Stores (optional, uncomment as needed)
# faiss-cpu>=1.7.4
# pymilvus>=2.3.0
